Add support for VMs with PVCs instead of DataVolume.
VMs that got restored from snapshots and previously
contained online added disks contain PVCs instead
of DataVolumes. These need special treatment as the
PVCs need to be re-created as well.

This commit also improves the check if the base OS
image for a DataVolume still exist. The oldest
image should only get picked if the source PVC
specified in the DataVolume is no longer present.

Finally the output is improved to be better
structured and readable.

Signed-off-by: hseipp <20093686+hseipp@users.noreply.github.com>